created EmployeeRepository and auto-wired service and emprepo

+++ emscrud/src/main/java/com/emscrud/emscrud/service/EmployeeService.java
@@ -10,5 +10,5 @@ public interface EmployeeService {
 
     public List<Employee> getEmployees();
 
-    public String deleteEmployee(int id);
+    public String deleteEmployee(Long id);
 }
